pyproject.toml
[tool.poetry]
name = "mercari"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyinstaller = "^4.5"
mercari-python = "^0.4"
PySimpleGUI = "^4.45.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
